Refactor version handling and add changelog context builder
