--- tools/release/files/__init__.py
@@ -1,24 +1,0 @@
-from .version_file import read_version_file, write_version_file
-from .meson import read_meson_version, write_meson_version
-from .package_json import read_package_json_version, write_package_json_version
-from .debian import (
-    DebianVersion,
-    read_debian_version,
-    write_debian_version,
-    format_debian_version,
-    parse_debian_version,
-)
-
-__all__ = [
-    "DebianVersion",
-    "read_version_file",
-    "write_version_file",
-    "read_meson_version",
-    "write_meson_version",
-    "read_package_json_version",
-    "write_package_json_version",
-    "read_debian_version",
-    "write_debian_version",
-    "format_debian_version",
-    "parse_debian_version",
-]
